helmet_detector/requirements.txt
tensorflow==2.10.0
# For GPU support, please use:
# tensorflow-gpu==2.10.0
tensorflow_io
tf_slim
numpy==1.23.5
pillow
lxml
Cython
contextlib2
pycocotools
matplotlib
tqdm
pandas
absl-py
protobuf==3.19.6
opencv-contrib-python
imutils
